new skill and status effect
new skill and status effect

+++ Assets/Scripts/Entity.cs
@@ -8,7 +8,7 @@ public class Entity : MonoBehaviour
     
     public struct Condition
     {
-        public int damage;  // If this condition deals damage, this is how much damage is dealt each time damage is dealt
+        public float damage;  // If this condition deals damage, this is how much damage is dealt each time damage is dealt
         public float damageTickRate; // The rate at which this condition will deal its damage
         public float duration; //The duration of the condition
         public ConditionEffect conditionType; //Name of condition
@@ -43,7 +43,7 @@ public class Entity : MonoBehaviour
             effectivePercent = Effectiveness;
         }
 
-        public Condition(float _duration, ConditionEffect _conditionType, int _damage, float _damageTickRate)
+        public Condition(float _duration, ConditionEffect _conditionType, float _damage, float _damageTickRate)
         {
             damage = _damage;
             damageTickRate = _damageTickRate;
@@ -55,7 +55,7 @@ public class Entity : MonoBehaviour
             effectivePercent = 1.0f;
         }
 
-        public Condition(float _duration, ConditionEffect _conditionType, float _effectivePercent, int _damage, float _damageTickRate)
+        public Condition(float _duration, ConditionEffect _conditionType, float _effectivePercent, float _damage, float _damageTickRate)
         {
             damage = _damage;
             damageTickRate = _damageTickRate;
@@ -258,7 +258,7 @@ public class Entity : MonoBehaviour
                         // After every second deal burn damage
                         if (condition.timePassed >= condition.damageTickRate)
                         {
-                            TakeDamage(condition.damage);
+                            TakeDamage((int)condition.damage);
 
                             // Reset the timePassed value
                             condition.ResetTimePassed();
@@ -286,7 +286,7 @@ public class Entity : MonoBehaviour
 
                         if (condition.timePassed >= condition.damageTickRate)
                         {
-                            TakeDamage(condition.damage);
+                            TakeDamage((int)condition.damage);
                             condition.ResetTimePassed();
                         }
                     }
